修复完成后的文件显示问题
